支持程序本身上传升级包更新并重启。

+++ rebootServer.sh
@@ -3,13 +3,16 @@
 LANG="zh_CN.UTF-8"
 export LANG
 
-pid=`ps -ef|grep "web-autoremote"|grep -v grep|grep -v PPID|grep -v tail|awk '{print $2}'`
+pid=`ps -ef|grep "web-autoremote"|grep -v grep|grep -v PPID|grep -v tail|grep -v "rebootServer.sh"|awk '{print $2}'`
 echo "pid:$pid"
 
-kill -9 $pid
-echo "kill -9 $pid"
+if [ $pid != '' ]
+then
+    kill -9 $pid
+    echo "kill -9 $pid"
+fi
 
-echo "nohup bash start.sh $1"
-nohup bash start.sh $1 &
+echo "bash start.sh $1"
+bash start.sh $1
 
 echo "rebootServer is OK!"
